# Optimization: Use userMetadata instead of individual lookups (non admin)